fix: Word 패키지 변경 및 require로 validation
